src/main/java/io/ljunggren/unit/converter/speed/base/BaseSpeedSiToUs.java
package io.ljunggren.unit.converter.speed.base;

import java.util.function.Function;

import io.ljunggren.unit.converter.Constants;
import io.ljunggren.unit.converter.Unit;
import io.ljunggren.unit.converter.base.BaseChain;

public class BaseSpeedSiToUs extends BaseChain {
    
    // Meter per second to Foot per second

    @Override
    public Function<Double, Double> baseUnit(Unit from, Unit to) {
        if (isSI(from) && isUS(to)) {
            return x -> x * Constants.METER_TO_FOOT;
        }
        return nextChain.baseUnit(from, to);
    }

}
